make file + norm some funcs

+++ ft_strncmp.c
@@ -9,10 +9,11 @@
 /*   Updated: 2023/01/18 19:35:27 by oloshcha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-int ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] != '\0' && i < n)
